[Rename] testbase -> base
